refactor(dot11): extract Dot11SequenceControl into sequence_control module
